requirements.txt
#---
pillow
xml
#---

tqdm
# --
torch
torchvision
# -- 
PyYAML==6.0.1